FMCAD2020: add paper scripts
This commit adds the scripts used in the FMCAD paper to a repository
where they can be accessed publicly.

The containing scripts allow to compile Xen 4.8.0 with CBMC, and show
case how XSAs can be analyzed with CPROVER tools to aid generating
security tests and analyzing the impact of an XSA. Check the README file
for more details.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>